requirements.txt
-i https://pypi.org/simple
aiomysql==0.2.0; python_version >= '3.7'
aiosqlite==0.20.0; python_version >= '3.8'
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.6.2.post1; python_version >= '3.9'
asyncmy==0.2.9; python_version >= '3.7' and python_version < '4.0'
cffi==1.17.1; python_version >= '3.8'
click==8.1.7; python_version >= '3.7'
cryptography==44.0.0; python_version >= '3.7' and python_full_version not in '3.9.0, 3.9.1'
exceptiongroup==1.2.2; python_version >= '3.7'
fastapi==0.115.5; python_version >= '3.8'
h11==0.14.0; python_version >= '3.7'
httptools==0.6.4; python_full_version >= '3.8.0'
idna==3.10; python_version >= '3.6'
iso8601==2.1.0; python_version >= '3.7' and python_version < '4.0'
pycparser==2.22; python_version >= '3.8'
pydantic==2.10.2; python_version >= '3.8'
pydantic-core==2.27.1; python_version >= '3.8'
pymysql==1.1.1; python_version >= '3.7'
pypika-tortoise==0.3.1; python_version >= '3.8' and python_version < '4.0'
python-dotenv==1.0.1; python_version >= '3.8'
pytz==2024.2
pyyaml==6.0.2; python_version >= '3.8'
sniffio==1.3.1; python_version >= '3.7'
starlette==0.41.3; python_version >= '3.8'
tortoise-orm[asyncmy]==0.22.1; python_version >= '3.8' and python_version < '4.0'
typing-extensions==4.12.2; python_version >= '3.8'
uvicorn[standard]==0.32.1; python_version >= '3.8'
uvloop==0.21.0; python_full_version >= '3.8.0'
watchfiles==1.0.0; python_version >= '3.9'
websockets==14.1; python_version >= '3.9'
